==> Tree_classifier_for_graphs/new_stuff/process/denonimator/04_12_loop_inference_2.ipynb <==
import pandas as pd

model_file = "/Users/rezadoobary/Downloads/bayes_results_merged/20251207-013645_train_567891011_notest_vstep1/merged_looporder_20251207-013650/inference_model_sklearn.pkl"
feature_names = '/Users/rezadoobary/Downloads/bayes_results_merged/20251207-013645_train_567891011_notest_vstep1/merged_looporder_20251207-013650/feature_names.json'

for i in range(5,10):   
    print(i,pd.read_csv(f"/Users/rezadoobary/Documents/MLCORRELATORS/ML-correlator/Tree_classifier_for_graphs/new_stuff/features/merged/{i}loops.csv")['Basic_num_nodes'].unique())

import json
import pickle

# Load modela
with open(model_file, "rb") as f:
    model = pickle.load(f)

# Load feature names
with open(feature_names, "r") as f:
    feature_names = json.load(f)

print("Model loaded:", type(model))
print("Number of features:", len(feature_names))


# test example


def correct_dataset(df):
       n = 16
       df["Centrality_closeness_mean_norm"] = df["Centrality_closeness_mean"] * (n - 1)
       df["Centrality_closeness_max_norm"]  = df["Centrality_closeness_max"]  * (n - 1)

       # Optionally handle rows where any inputs are NaN:
       mask =  ~np.isfinite(df["Centrality_closeness_mean"]) | \
              ~np.isfinite(df["Centrality_closeness_max"])

       df.loc[mask, ["Centrality_closeness_mean_norm", "Centrality_closeness_max_norm"]] = np.nan
       return df


from tqdm import tqdm
import numpy as np

def predict_with_progress(model, df, feature_names, batch_size=5000):
    X = df[feature_names]
    n = len(X)
    preds = []

    for start in tqdm(range(0, n, batch_size), desc="Predicting", unit="rows"):
        end = min(start + batch_size, n)
        batch = X.iloc[start:end]
        batch_pred = model.predict_proba(batch)[:, 1]
        preds.append(batch_pred)

    return np.concatenate(preds)

# Run with progress

scores_12loops = []
for i in range(1,21):
    df = pd.read_parquet(f"/Users/rezadoobary/Documents/MLCORRELATORS/ML-correlator/Tree_classifier_for_graphs/new_stuff/features/12loops/merged_data/{i}part.parquet")
    df = correct_dataset(df)
    predictions = predict_with_progress(model, df, feature_names)
    scores_12loops.extend(predictions)
    print(i, len(scores_12loops))


scores_12loops = [float(x) for x in scores_12loops]
pickle.dump(
    scores_12loops,
    open('/Users/rezadoobary/Documents/MLCORRELATORS/ML-correlator/Tree_classifier_for_graphs/new_stuff/results/scores/loop_scores_vs1.pkl', 'wb')
)

# loop order    

len(coefficients)

from tqdm import tqdm
import numpy as np



coefficients = []
for i in range(1,21):
    df = pd.read_parquet(f"/Users/rezadoobary/Documents/MLCORRELATORS/ML-correlator/Tree_classifier_for_graphs/new_stuff/features/12loops/merged_data/{i}part.parquet")
    coefficients.extend(df['COEFFICIENTS'].values)

coefficients = [int(x) for x in coefficients]

pickle.dump(
    coefficients,
    open('/Users/rezadoobary/Documents/MLCORRELATORS/ML-correlator/Tree_classifier_for_graphs/new_stuff/results/scores/coefficients.pkl', 'wb')
)

from sklearn.metrics import roc_auc_score

coefficients = pickle.load(
    open('/Users/rezadoobary/Documents/MLCORRELATORS/ML-correlator/Tree_classifier_for_graphs/new_stuff/results/scores/coefficients.pkl', 'rb')
)

loop_scores = pickle.load(
    open('/Users/rezadoobary/Documents/MLCORRELATORS/ML-correlator/Tree_classifier_for_graphs/new_stuff/results/scores/loop_scores_vs1.pkl', 'rb')
)


roc_auc_score(coefficients, loop_scores)

model_file = "/Users/rezadoobary/Downloads/bayes_results_merged/20251206-205803_train_5678910_test11_vstep1/merged_looporder_20251206-205809/inference_model_sklearn.pkl"
feature_names = '/Users/rezadoobary/Downloads/bayes_results_merged/20251206-205803_train_5678910_test11_vstep1/merged_looporder_20251206-205809/feature_names.json'

import json
import pickle

# Load modela
with open(model_file, "rb") as f:
    model = pickle.load(f)

# Load feature names
with open(feature_names, "r") as f:
    feature_names = json.load(f)

print("Model loaded:", type(model))
print("Number of features:", len(feature_names))


from tqdm import tqdm
import numpy as np

def predict_with_progress(model, df, feature_names, batch_size=5000):
    X = df[feature_names]
    n = len(X)
    preds = []

    for start in tqdm(range(0, n, batch_size), desc="Predicting", unit="rows"):
        end = min(start + batch_size, n)
        batch = X.iloc[start:end]
        batch_pred = model.predict_proba(batch)[:, 1]
        preds.append(batch_pred)

    return np.concatenate(preds)

# Run with progress

scores_12loops = []
for i in range(1,21):
    df = pd.read_parquet(f"/Users/rezadoobary/Documents/MLCORRELATORS/ML-correlator/Tree_classifier_for_graphs/new_stuff/features/12loops/merged_data/{i}part.parquet")
    df = correct_dataset(df)
    predictions = predict_with_progress(model, df, feature_names)
    scores_12loops.extend(predictions)
    print(i, len(scores_12loops))


scores_12loops = [float(x) for x in scores_12loops]
pickle.dump(
    scores_12loops,
    open('/Users/rezadoobary/Documents/MLCORRELATORS/ML-correlator/Tree_classifier_for_graphs/new_stuff/results/scores/loop_scores_vs1_l2.pkl', 'wb')
)

scores_12loops = [float(x) for x in scores_12loops]
roc_auc_score(coefficients, scores_12loops)

vals = np.array(np.array(strat_scores) > 0.29, dtype = int)

from sklearn.metrics import precision_score, recall_score, f1_score, balanced_accuracy_score
import pandas as pd

# Example: y_true and y_pred already defined
y_true = coefficients
y_pred = vals

metrics = {
    "Precision": precision_score(y_true, y_pred, zero_division=0),
    "Recall": recall_score(y_true, y_pred, zero_division=0),
    "F1 Score": f1_score(y_true, y_pred, zero_division=0),
    "Balanced Accuracy": balanced_accuracy_score(y_true, y_pred)
}

metrics_df = pd.DataFrame(metrics, index=["Score"]).T  # T = Transpose for column format
metrics_df


vals = np.array(np.array(loop_scores) > 0.38, dtype = int)

from sklearn.metrics import precision_score, recall_score, f1_score, balanced_accuracy_score
import pandas as pd

# Example: y_true and y_pred already defined
y_true = coefficients
y_pred = vals

metrics = {
    "Precision": precision_score(y_true, y_pred, zero_division=0),
    "Recall": recall_score(y_true, y_pred, zero_division=0),
    "F1 Score": f1_score(y_true, y_pred, zero_division=0),
    "Balanced Accuracy": balanced_accuracy_score(y_true, y_pred)
}

metrics_df = pd.DataFrame(metrics, index=["Score"]).T  # T = Transpose for column format
metrics_df


# quikc look at f-graphs

import pandas as pd
from fgraph_features_cli3 import build_graphs_from_row, extract_features_row, parse_groups

df = pd.read_csv("/Users/rezadoobary/Downloads/graph_data_5.csv")
row = df.iloc[0]

groups = parse_groups("all")

feats = extract_features_row(
    row,
    groups=groups,
    lap_eigs_k=10,
    netlsd_points=32,
    netlsd_tmin=0.01,
    netlsd_tmax=100.0,
    ind4_exact_nmax=50,
    ind4_max_samples=50000,
    ind5_sample_frac=1.0,
    distance_max_sources=200,
    distance_exact_if_leq=500,
    seed=42,
    intra_workers=0,
    betweenness_k=0,
)

for k in sorted([c for c in feats.keys() if not str(feats[c]) == "nan"])[:40]:
    print(k, "=", feats[k])

fgraph5 = pd.read_csv("/Users/rezadoobary/Documents/MLCORRELATORS/ML-correlator/Tree_classifier_for_graphs/new_stuff/features/fgraphs2/5_fgraph_feats.csv")

fgraph5['COEFFICIENTS']
